+++ src/components/Button/index.js
@@ -1,4 +1,4 @@
 export { default as Button } from './Button/Button';
-export { default as ButtonToggle } from './ButtonToggle/ButtonToggle';
+// export { default as ButtonToggle } from './ButtonToggle/ButtonToggle';
 export { default as ButtonArrow } from './ButtonArrow/ButtonArrow';
 export { default as ButtonPlus } from './ButtonPlus/ButtonPlus';
