+++ src/pages/CardPostListPage/index.js
@@ -1,0 +1,1 @@
+export { default as CardPostListPage } from './CardPostListPage';
