--- src/pages/LandingPage.jsx
@@ -1,11 +1,0 @@
-import Header from '../components/Header/Header.jsx';
-
-function LandingPage() {
-  return (
-    <>
-      <Header />
-    </>
-  );
-}
-
-export default LandingPage;
