+++ src/pages/LoadingPage/index.js
@@ -1,0 +1,1 @@
+export { default as LoadingPage } from './LoadingPage';
