+++ src/pages/PostMessagePage/index.js
@@ -1,0 +1,1 @@
+export { default as PostMessagePage } from './PostMessagePage';
