--- src/pages/index.js
@@ -1,1 +1,0 @@
-// init
