+++ src/pages/listpage/index.js
@@ -1,0 +1,1 @@
+export { default as ListPage } from './ListPage';
