--- src/pages/listpage/index.jsx
@@ -1,9 +1,0 @@
-function ListPage() {
-  return (
-    <>
-      <h1>List Page!</h1>
-    </>
-  );
-}
-
-export default ListPage;
