+++ src/pages/postpage/index.js
@@ -1,0 +1,1 @@
+export { default as PostPage } from './PostPage';
